debugging preprocessor for preprocessing a text without given labels

+++ hotam/utils.py
@@ -239,7 +239,7 @@ def dynamic_update(src, v):
     a = np.array(list(src.shape[1:]))
     b = np.array(list(v.shape))
     new_shape = np.maximum(a, b)
-    new_src = np.zeros((src.shape[0]+1, *new_shape))
+    new_src = np.zeros((src.shape[0]+1, *new_shape), dtype=src.dtype)
 
     if len(v.shape) > 2:
         new_src[:src.shape[0],:src.shape[1], :src.shape[2]] = src
